requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-ckeditor==6.7.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.13
idna==3.10
phonenumbers==9.0.1
pillow==11.1.0
Pygments==2.19.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
validators==0.34.0
